requirements.txt
Flask-SQLAlchemy
Flask-Cors
python-decouple
Flask-RESTful
ftplib